% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwDiff.R
\name{pairwDiff}
\alias{pairwDiff}
\title{Pairwise Difference}
\usage{
pairwDiff(SFS)
}
\arguments{
\item{SFS}{vector with the site frequency spectrum.}
}
\value{
The pairwise difference.
}
\description{
The function calculates the pairwise difference based on the site
frequency spectrum.
}
\details{
The site frequency spectrum is a vector of length \eqn{n-1}, where
\eqn{n} is the sample size. The \eqn{i}'th entry is the number of
segregating sites where a mutation occur in \eqn{i} sequences, and
thus all entries must be natural numbers (0 included).
If one runs \code{pairwDiff}
with a vector that contains anything other than natural numbers,
an error will occur.

For details about the pairwise difference, consult the analyzeDNA vignette.
}
\examples{
pairwDiff(c(2,1,0,0))

}
\references{
Wakeley J. (2009) \emph{Coalescent Theory: An Introduction}. Colorado:
Roberts and Company Publishers.
}
