% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DNAmat_class.R
\name{plot.DNAmat_class}
\alias{plot.DNAmat_class}
\title{plot.DNAmat_class}
\usage{
\method{plot}{DNAmat_class}(DNAmat, ...)
}
\arguments{
\item{DNAmat}{matrix; either a single-nucleotide polymorphism matrix or a segregating
sites matrix. See \code{\link{SNP}} for explanation.}
}
\description{
Plots the site frequency spectrum of \code{DNAmat}.
}
\examples{
plot(simDNAseq(n = 8, seqLen = 20, mutRate = 5, popType = "varPop",
expRate = 1.5), col="red")

SegSitesMat <- matrix(sample(x=c(0, 1), size=26*4, replace=TRUE, prob=c(3/4, 1/4)),
nrow=4, ncol=26, byrow=TRUE)
plot.DNAmat_class(SegSitesMat)

}
