% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_SFS_class.R
\name{plot.SFS_class}
\alias{plot.SFS_class}
\title{Site frequency spectrum plot}
\usage{
\method{plot}{SFS_class}(SFS, xlab = "Index i",
  ylab = "Number of i-tons", main = "Site frequency spectrum", ...)
}
\arguments{
\item{SFS}{vector with the site frequency spectrum of class 'SFS_class'.}

\item{xlab}{a title for the x axis. Is by default "Index i". See \code{\link[graphics]{title}}.}

\item{ylab}{a title for the y axis. Is by default "Number of i-tons". See \code{\link[graphics]{title}}.}

\item{main}{an overall title for the plot. Is by default "Site frequency spectrum". See \code{\link[graphics]{title}}.}

\item{...}{arguments passed on from \code{\link[graphics]{plot}}.}
}
\description{
Plots the site frequency spectrum.
}
\details{
The terminology "Number of i-tons" corresponds to the i'th entry in the site
frequency spectrum. See \code{\link{SFS}} for further explanation.
}
\examples{
# Using the functions simDNAseq and SFS:
DNAmat <- simDNAseq(n = 24, seqLen = 58, mutRate = 7,
                    popType = "sudExpPop", expansionTime = 3,
                    proportion = 0.6)
plot(SFS(DNAmat), col="green")


# Creating the SFS vector by hand:
SFSvec <- c(2, 4, 1, 0, 0)
class(SFSvec) <- "SFS_class"
plot(SFSvec)

}
