% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_population.R
\name{print.population}
\alias{print.population}
\title{Print DNA sequences}
\usage{
## S3 method for class 'population'
print.population(x, ...)
}
\arguments{
\item{x}{a segregating sites matrix. See details.}
}
\description{
Prints a segregating sites matrix.
}
\details{
A segregating sites matrix consists of zeroes and ones.
Here the number 1 indicates that a mutation occured at this specific site, and
the number 0 indicates that no mutation occured.

It is used for printing the simulated segregating sites matrix from \code{simDNAseq}.
}
\examples{
SegMat <- matrix(c(1, 1, 1, 1, 0,
                   0, 1, 1, 1, 0,
                   0, 0, 0, 1, 0,
                   0, 0, 0, 0, 1,
                   0, 0, 0, 0, 0), byrow=T, nrow=5)
class(SegMat) <- "population"
print(SegMat)

}
