% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_SFS_class.R
\name{print.SFS_class}
\alias{print.SFS_class}
\title{Print site frequency spectrum}
\usage{
\method{print}{SFS_class}(SFS, ...)
}
\arguments{
\item{SFS}{a vector with the site frequency spectrum.}

\item{...}{arguments passed on from \code{\link[base]{print}}.}
}
\description{
Prints a vector with the site frequency spectrum.
}
\details{
It is used for printing the site frequency spectrum from \code{SFS}.
}
\examples{
SFSvec <- c(4, 1, 2, 0, 0)
class(SFSvec) <- "SFS_class"
print(SFSvec)

}
