% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFS.R
\name{SFS}
\alias{SFS}
\title{SFS}
\usage{
SFS(SNP)
}
\description{
Calculte the site frequency spectrum
}
\details{
The function calculates the site frequency spectrum and outputs it as a vector. That is, the i'th entry of the output vector is the number of mutations that happened on a branch where exactly i sequences has coalesced.
}
