% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP.R
\name{SNP}
\alias{SNP}
\title{SNP (Single-Nucleotide Polymorphism)}
\usage{
SNP(matrix(c(0,0,0,1,0,0,0,0,1),nrow=3))
}
\arguments{
\item{DNAseq}{A matrix with DNA sequences. Each row is a DNA sequence represented by 0's and 1's, where 1 indicates that a mutation has happened,0 that no mutations has happened.}
}
\value{
The output of the function is a list.
The first element in the list is a vector containing the positions on which one or more mutations has occured.
The second element is the SNP matrix itself.
}
\description{
The function calculates the SNP matrix.
}
\examples{
SNP(matrix(c(0,0,0,1,0,0,0,0,1),nrow=3))
## $positions
## [1] 2 3
##
## $SNPmat
##      [,1] [,2]
## [1,]    1    0
## [2,]    0    0
## [3,]    0    1

}
