% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutRate.R
\name{mutRate}
\alias{mutRate}
\title{Mutation Rate}
\usage{
mutRate(SFS)
}
\arguments{
\item{SFS}{vector with the site frequency spectrum.}
}
\value{
List containing the following components:
\item{Watterson}{Watterson's estimator.}
\item{pairwDiff}{The pairwise difference estimator.}
}
\description{
The function calculates two different estimates of the mutation rate:
Watterson's estimator and the pairwise difference estimator.
}
\details{
See \code{\link{SFS}} for explanation of the site frequency spectrum.

If one runs \code{mutRate}
with a vector that contains anything other than natural numbers,
an error will occur.

For details about Watterson's estimator and the pairwise difference
estimator, consult the analyzeDNA vignette.
}
\examples{
mutRate(c(2,1,0,0,0,1,0))

}
\references{
Wakeley J. (2009) \emph{Coalescent Theory: An Introduction}. Colorado:
Roberts and Company Publishers.
}
