% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutRate.R
\name{mutRate}
\alias{mutRate}
\title{Mutation Rate}
\usage{
mutRate(c(2,1,0,0,0))
}
\arguments{
\item{SFS}{The input is a vector with the site frequency spectrum.}
}
\value{
The function returns a list with Watterson's estimator and the pairwise difference estimator
}
\description{
The function calculates two different estimates of the mutation rate: Watterson's estimator and the pairwise difference estimator.
}
\examples{
mutRate(c(2,1,0,0,0,1,0))
## $Watterson
## [1] 1.5427
##
## $pairwDiff
## [1] 1.357143

}
