% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwDiff.R
\name{pairwDiff}
\alias{pairwDiff}
\title{Pairwise Difference}
\usage{
pairwDiff(c(2,1,0,0))
}
\arguments{
\item{SFS}{The input is a vector with the site frequency spectrum.}
}
\value{
The function returns the pairwise difference.
}
\description{
The function calculates the pairwise difference based on the site frequency spectrum.
}
\examples{
pairwDiff(c(2,1,0,0))
## [1] 1.4

}
