% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DNAmat_class.R
\name{plot.DNAmat_class}
\alias{plot.DNAmat_class}
\title{Site frequency spectrum plot}
\usage{
## S3 method for class 'DNAmat_class'
plot(DNAmat, xlab="Index i", ylab="Number of i-tons",
main="Site frequency spectrum", ...)
}
\arguments{
\item{DNAmat}{matrix; either a single-nucleotide polymorphism matrix or a segregating
sites matrix. See \code{\link{SNP}} for explanation.}
}
\description{
Plots the site frequency spectrum of \code{DNAmat}.
}
\examples{
plot(simDNAseq(n = 8, seqLen = 20, mutRate = 5,
               popType = "varPop", expRate = 1.5), col="red")


SegSitesMat <- matrix(sample(x=c(0, 1), size=26*12,
                      replace=TRUE, prob=c(0.9, 0.1)),
                      nrow=12, ncol=26, byrow=TRUE)
plot.DNAmat_class(SegSitesMat)


SNPmat <- matrix(c(1, 0, 0, 1, 1, 0, 0,
                   0, 1, 0, 1, 0, 0, 0,
                   0, 0, 1, 0, 0, 1, 1,
                   1, 0, 0, 0, 0, 0, 1,
                   1, 0, 0, 1, 0, 0, 0,
                   1, 0, 1, 1, 0, 1, 0), nrow=6, ncol=7, byrow=TRUE)
class(SNPmat) <- "DNAmat_class"
plot(SNPmat, main="SFS of SNP matrix")

}
