% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFS.R
\name{SFS}
\alias{SFS}
\title{Site Frequency Spectrum}
\usage{
SFS(DNAmat)
}
\arguments{
\item{DNAmat}{matrix; either a single-nucleotide polymorphism matrix or a segregating
sites matrix. See details.}
}
\value{
The site frequency spectrum.
}
\description{
The function calculates the site frequency spectrum.
}
\details{
The site frequency spectrum is a vector of length \eqn{n-1}, where
\eqn{n} is the sample size. The \eqn{i}'th entry is the number of
segregating sites where a mutation occur in \eqn{i} sequences.

The input to the function should be either a single-nucleotide
polymorphism matrix or a segregating sites matrix. Either way,
each row in the matrix represents a DNA sequence.
The matrix should have a 1 where a mutation has occured and a 0 otherwise.
If this is not the case, it will cause an error.
If there exists a position on which all sequences have a mutation,
it will cause a warning, and such a position will be treated
as if it had no mutations.
}
\examples{
SFS(matrix(c(0,1,0,1,0,0,0,0,0),nrow=3))

}
