% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFS.R
\name{SFS}
\alias{SFS}
\title{Site Frequency Spectrum}
\usage{
SFS(matrix(c(0,1,0,1,0,0,0,0,0),nrow=3))
}
\arguments{
\item{SNP}{The input is a SNP (single-nucleotide polymorphism) matrix, or a matrix with DNA sequences, where each row is a sequence. The matrix should have a 1 where a mutation has occured and a 0 otherwise. If this is not the case, it will cause an error. Also, an error will be caused if there exists a position on which all sequences have a mutation, as this is an indication that the ancestral state is in fact not the ancestral state.}
}
\value{
The function returns a vector with the site frequency spectrum
}
\description{
The function SFS calcultes the site frequency spectrum.
}
\details{
The function calculates the site frequency spectrum and outputs it as a vector. That is, the i'th entry of the output vector is the number of mutations that happened on a branch where exactly i sequences has coalesced.
}
\examples{
SFS(matrix(c(0,1,0,1,0,0,0,0,0),nrow=3))
## [1] 2 0

}
