% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDNA.R
\docType{package}
\name{simDNA}
\alias{simDNA}
\alias{simDNA-package}
\title{The simDNA package}
\description{
R functions for simulating and analyzing DNA sequences.
}
\details{
Contains function to simulate segregating sites matrices for various population types.
It also contains functions to analyze DNA based on segregating sites matrices or the site
frequency spectrum.

The names of these functions are the following
\itemize{
   \item \code{\link{mutRate}}
   \item \code{\link{pairwDiff}}
   \item \code{\link{print.DNAmat_class}}
   \item \code{\link{SFS}}
   \item \code{\link{simDNAseq}}
   \item \code{\link{SNP}}
   \item \code{\link{TajimaD}}
}
}
\author{
Elias Arildskov, Camilla P. Kudahl and Maiken B. Poulsen.
}
