% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP.R
\name{SNP}
\alias{SNP}
\title{SNP (Single-Nucleotide Polymorphism) matrix}
\usage{
SNP(segSites)
}
\arguments{
\item{segSites}{segregating sites matrix. See details.}
}
\value{
List containing the following components:
\item{positions}{the positions on which one or more mutations occurred.}
\item{SNPmat}{the single-nucleotide polymorphism matrix with class 'DNAmat_class'.}
}
\description{
The function calculates the SNP matrix.
}
\details{
In the segregating sites matrix, each row is a DNA sequence represented
by zeros and ones, where 1 indicates that a mutation has occurred,
0 that no mutations have occurred. If the \code{segSites} matrix contains
anything other than zeros and ones, it will cause an error.

The SNP matrix extracts the columns from the segregating sites matrix
where one or more mutations have happened, and thus it is a more compact
representation than the segregating sites matrix.
}
\examples{
SNP(matrix(c(0,0,0,1,0,0,0,0,1),nrow=3))

}
