% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_DNAmat_class.R
\name{summary.DNAmat_class}
\alias{summary.DNAmat_class}
\title{Summary statistics}
\usage{
## S3 method for class DNAmat_class
summary(DNAmat)
}
\arguments{
\item{DNAmat}{matrix; either a single-nucleotide polymorphism matrix or a segregating
sites matrix.}
}
\description{
Gives summary statistics useful for analyzing DNA.
}
\examples{
DNAmat <- matrix(c(0,1,0,0,
                   0,1,0,1,
                   0,0,0,0,
                   0,0,0,0), 4,4, byrow=TRUE)
class(DNAmat) <- "DNAmat_class"
summary(DNAmat)

}
\seealso{
\code{\link{SFS}}, \code{\link{mutRate}}, \code{\link{TajimaD}}
}
