% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_SFS_class.R
\name{summary.SFS_class}
\alias{summary.SFS_class}
\title{Summary statistics}
\usage{
## S3 method for class SFS_class
summary(SFS)
}
\arguments{
\item{SFS}{vector with the site frequency spectrum.}
}
\description{
Gives summary statistics useful for analyzing DNA.
}
\examples{
SFS <- c(2,1,0,0,0,1,0,0)
class(SFS) <- "SFS_class"
summary(SFS)

}
\seealso{
\code{\link{mutRate}}, \code{\link{TajimaD}}
}
