% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_SFS_class.R
\name{summary.SFS_class}
\alias{summary.SFS_class}
\title{Summary statistics}
\usage{
## S3 method for class 'SFS_class'
summary(SFS)
}
\arguments{
\item{SFS}{vector with the site frequency spectrum of class 'SFS_class'.}
}
\value{
List containing the following components:
\item{Watterson}{Watterson's estimator.}
\item{pairwDiff}{The pairwise difference estimator.}
\item{TajimaD}{Tajima's D}
}
\description{
Calculates summary statistics useful for analyzing DNA.
}
\examples{
# Using the function SFS:
DNAmat <- matrix(c(0,1,0,0,
                   0,1,0,1,
                   0,0,0,0,
                   0,0,0,0), 4,4, byrow=TRUE)
summary(SFS(DNAmat))

# Creating SFS by hand:
SFS <- c(2,1,0,0,0,1,0,0)
class(SFS) <- "SFS_class"
summary(SFS)

}
\seealso{
\code{\link{mutRate}}, \code{\link{TajimaD}}
}
