% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TajimaD.R
\name{TajimaD}
\alias{TajimaD}
\title{Tajima's D}
\usage{
TajimaD(c(3,0,2,1,0,0,0))
}
\arguments{
\item{SFS}{The input is a vector with the site frequency spectrum.}
}
\value{
The function returns Tajima's D.
}
\description{
The function calculates Tajima's D.
}
\examples{
TajimaD(c(3,0,2,1,0,0,0))

}
