% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TajimaD.R
\name{TajimaD}
\alias{TajimaD}
\title{Tajima's D}
\usage{
TajimaD(SFS)
}
\arguments{
\item{SFS}{vector with the site frequency spectrum.}
}
\value{
Tajima's D.
}
\description{
The function calculates Tajima's D.
}
\details{
See \code{\link{SFS}} for explanation of the site frequency spectrum.

If one runs \code{TajimaD}
with a vector that contains anything other than natural numbers,
an error will occur.

For details about Tajima's D, consult the analyzeDNA vignette.
}
\examples{
TajimaD(c(3,0,2,1,0,0,0))

}
\references{
Wakeley J. (2009) \emph{Coalescent Theory: An Introduction}. Colorado:
Roberts and Company Publishers.
}
